fixed all for master

+++ src/main/java/com/example/java/test/junior/developer/service/CategoryService.java
@@ -16,6 +16,7 @@ import org.springframework.transaction.annotation.Transactional;
 public class CategoryService {
 
   private final CategoryRepository categoryRepository;
+
   private final CategoryMapper categoryMapper;
 
   @Transactional
@@ -34,7 +35,7 @@ public class CategoryService {
   }
 
   @Transactional
-  public CategoryDto updateCategory(int id, CategoryDto dto) {
+  public CategoryDto updateCategory(Long id, CategoryDto dto) {
     final var category = categoryMapper.toEntity(dto);
     category.setId(id);
     final var saved = categoryRepository.save(category);
@@ -42,7 +43,9 @@ public class CategoryService {
   }
 
   @Transactional
-  public void deleteCategory(int id) {
+  public void deleteCategory(Long id) {
     categoryRepository.deleteById(id);
   }
+
 }
+
